Refactor code base
Merged atomicChess.ts and validator.ts

Fixed imports

Renamed CastleType enum to CastleSide
