+++ src/atomic-chess/atomicChess.ts
@@ -4,6 +4,12 @@ export enum Color {
   BLACK = 'black',
 }
 
+// Enumerate the possible types of castling
+export enum CastleSide {
+  KINGSIDE = 'kingside',
+  QUEENSIDE = 'queenside',
+}
+
 // Enumerate the possible types of moves
 export enum MoveType {
   CAPTURE = 'standard capture',
@@ -13,18 +19,24 @@ export enum MoveType {
   STANDARD_MOVE = 'standard move',
 }
 
-// Enumerate the possible types of castling
-export enum CastleType {
-  KINGSIDE = 'kingside',
-  QUEENSIDE = 'queenside',
-}
-
 // Enumerate the possible types of results when the game ends
 export enum GameOverType {
   WHITE_WIN = 'w',
   BLACK_WIN = 'b',
   DRAW = 'd',
   STALEMATE = 's',
+}
+
+export enum Flag {
+  PROMOTION = 'promotion',
+  DOUBLE = 'double',
+  KING_MOVE = 'king move',
+  KINGSIDE_ROOK_MOVE = 'kingside rook move',
+  QUEENSIDE_ROOK_MOVE = 'queenside rook move',
+  KINGSIDE_ROOK_CAPTURED = 'kingside rook captured',
+  QUEENSIDE_ROOK_CAPTURED = 'queenside rook captured',
+  PAWN_MOVE = 'pawn move',
+  CAPTURE = 'capture'
 }
 
 // Enumerate the possible types of chess pieces
@@ -41,6 +53,20 @@ export enum PieceType {
 export interface Move {
   from: Square,
   to: Square,
+}
+
+export interface ExternalMove {
+  color: Color;
+  from: Square;
+  to: Square;
+  piece: Piece;
+  flags: Flag[];
+  result: Chessboard;
+  moveType: MoveType;
+  actions: {
+    move: Move;
+    explode: boolean;
+  }[];
 }
 
 // Define the structure for a castling move, including the movement of the king, rook, and squares between them.
@@ -69,7 +95,7 @@ export type Square =
 
 // Define the possible grid indices on a chessboard
 type Index = 0 | 1 | 2 | 3 | 4 | 5 | 6 | 7;
-export type SquareIndex = [Index, Index];
+export type SquareGridIndex = [Index, Index];
 
 // Define the possible grid offsets of each move on a chessboard
 type Offset = -7 | -6 | -5 | -4 | -3 | -2 | -1 | Index;
@@ -90,7 +116,7 @@ export type Piece =
 export type Chessboard = Record<Square, Piece | null>;
 
 // Define the structure for castling rights
-export type CastlingRights = Record<Color, Record<CastleType, boolean>>;
+export type CastlingRights = Record<Color, Record<CastleSide, boolean>>;
 
 // Define the structure for FEN (Forsyth–Edwards Notation) of the state of a chess game
 export type FEN = {
@@ -135,7 +161,7 @@ export const PIECE_TO_TYPE: Record<Piece, PieceType> = {
 };
 
 // Map chess squares to grid indices
-export const SQUARE_TO_INDEX: Record<Square, SquareIndex> = {
+export const SQUARE_TO_GRID_INDEX: Record<Square, SquareGridIndex> = {
   a8: [7, 0], b8: [7, 1], c8: [7, 2], d8: [7, 3], e8: [7, 4], f8: [7, 5], g8: [7, 6], h8: [7, 7],
   a7: [6, 0], b7: [6, 1], c7: [6, 2], d7: [6, 3], e7: [6, 4], f7: [6, 5], g7: [6, 6], h7: [6, 7],
   a6: [5, 0], b6: [5, 1], c6: [5, 2], d6: [5, 3], e6: [5, 4], f6: [5, 5], g6: [5, 6], h6: [5, 7],
@@ -148,46 +174,12 @@ export const SQUARE_TO_INDEX: Record<Square, SquareIndex> = {
 
 // Define possible movement patterns
 export const MOVE_PATTERNS: Record<string, SquareOffset[]> = {
-  ALL: [
-    [0, 1],
-    [0, -1],
-    [1, 0],
-    [-1, 0],
-    [1, 1],
-    [-1, 1],
-    [-1, -1],
-    [1, -1],
-  ],
-  DIAGONAL: [
-    [1, 1],
-    [-1, 1],
-    [-1, -1],
-    [1, -1],
-  ],
-  HORIZONTAL: [
-    [0, 1],
-    [0, -1],
-  ],
-  L_SHAPE: [
-    [1, 2],
-    [2, 1],
-    [2, -1],
-    [1, -2],
-    [-1, -2],
-    [-2, -1],
-    [-2, 1],
-    [-1, 2],
-  ],
-  PLUS_SHAPE: [
-    [0, 1],
-    [0, -1],
-    [1, 0],
-    [-1, 0],
-  ],
-  VERTICAL: [
-    [1, 0],
-    [-1, 0],
-  ],
+  ALL: [[0, 1], [0, -1], [1, 0], [-1, 0], [1, 1], [-1, 1], [-1, -1], [1, -1]],
+  DIAGONAL: [[1, 1], [-1, 1], [-1, -1], [1, -1]],
+  HORIZONTAL: [[0, 1], [0, -1]],
+  L_SHAPE: [[1, 2], [2, 1], [2, -1], [1, -2], [-1, -2], [-2, -1], [-2, 1], [-1, 2]],
+  PLUS_SHAPE: [[0, 1], [0, -1], [1, 0], [-1, 0]],
+  VERTICAL: [[1, 0], [-1, 0]],
 };
 
 // Map pieces to their standard movement patterns
@@ -225,28 +217,12 @@ export const PIECE_CAPTURE_PATTERNS: Record<Piece, MovePattern> = {
 // Map player colors to castles
 export const CASTLE_MOVES: Record<Color, { kingside: CastleMove, queenside: CastleMove }> = {
   [Color.WHITE]: {
-    kingside: {
-      kingMove: { from: 'e1', to: 'g1' },
-      rookMove: { from: 'h1', to: 'f1' },
-      between: ['f1', 'g1'],
-    },
-    queenside: {
-      kingMove: { from: 'e1', to: 'c1' },
-      rookMove: { from: 'a1', to: 'd1' },
-      between: ['d1', 'c1', 'b1'],
-    },
+    kingside: { kingMove: { from: 'e1', to: 'g1' }, rookMove: { from: 'h1', to: 'f1' }, between: ['f1', 'g1'] },
+    queenside: { kingMove: { from: 'e1', to: 'c1' }, rookMove: { from: 'a1', to: 'd1' }, between: ['d1', 'c1', 'b1'] },
   },
   [Color.BLACK]: {
-    kingside: {
-      kingMove: { from: 'e8', to: 'g8' },
-      rookMove: { from: 'h8', to: 'f8' },
-      between: ['f8', 'g8'],
-    },
-    queenside: {
-      kingMove: { from: 'e8', to: 'c8' },
-      rookMove: { from: 'a8', to: 'd8' },
-      between: ['d8', 'c8', 'b8'],
-    },
+    kingside: { kingMove: { from: 'e8', to: 'g8' }, rookMove: { from: 'h8', to: 'f8' }, between: ['f8', 'g8'] },
+    queenside: { kingMove: { from: 'e8', to: 'c8' }, rookMove: { from: 'a8', to: 'd8' }, between: ['d8', 'c8', 'b8'] },
   },
 };
 
@@ -286,7 +262,6 @@ export const PROMOTABLE_PIECES: Record<Color, PromotablePiece[]> = {
   [Color.BLACK]: ['q', 'n', 'r', 'b']
 };
 
-
 // Define initial chess position of a chess game
 export const INITIAL_CHESSBOARD_POSITION: Record<Square, Piece | null> = {
   a8: 'r', b8: 'n', c8: 'b', d8: 'q', e8: 'k', f8: 'b', g8: 'n', h8: 'r',
@@ -301,8 +276,8 @@ export const INITIAL_CHESSBOARD_POSITION: Record<Square, Piece | null> = {
 
 // Define initial castling rights of a chess game
 export const INITIAL_CASTLING_RIGHTS: CastlingRights = {
-  [Color.WHITE]: { [CastleType.KINGSIDE]: true, [CastleType.QUEENSIDE]: true },
-  [Color.BLACK]: { [CastleType.KINGSIDE]: true, [CastleType.QUEENSIDE]: true }
+  [Color.WHITE]: { [CastleSide.KINGSIDE]: true, [CastleSide.QUEENSIDE]: true },
+  [Color.BLACK]: { [CastleSide.KINGSIDE]: true, [CastleSide.QUEENSIDE]: true }
 };
 
 // Define FEN representation of initial state of a chess game
@@ -316,7 +291,7 @@ export const INITIAL_GAMESTATE: FEN = {
 };
 
 // Define array of all chessboard squares
-export const CHESSBOARD_SQUARES: Square[] = Object.keys(SQUARE_TO_INDEX) as Square[];
+export const CHESSBOARD_SQUARES: Square[] = Object.keys(SQUARE_TO_GRID_INDEX) as Square[];
 
 // Returns the opposite player color of the given color
 export function getEnemyColor(color: Color): Color {
@@ -327,16 +302,316 @@ export function getEnemyColor(color: Color): Color {
 export function gridIndexToSquare(index: Pos): Square | null {
   const [r, f] = index;
   if (r < 0 || r > 7 || f < 0 || f > 7) return null;
-  return squareIndexToSquare(index as SquareIndex);
+  return squareIndexToSquare(index as SquareGridIndex);
 }
 
 // Converts a square index to the corresponding square
-export function squareIndexToSquare([r, f]: SquareIndex): Square {
+export function squareIndexToSquare([r, f]: SquareGridIndex): Square {
   return 'abcdefgh'[f] + '12345678'[r] as Square;
 }
 
 // Moves a square by a given offset
 export function moveSquare(square: Square, [dr, dc]: SquareOffset, steps: number = 1): Square | null {
-  const [r, c] = SQUARE_TO_INDEX[square];
+  const [r, c] = SQUARE_TO_GRID_INDEX[square];
   return gridIndexToSquare([r + steps * dr, c + steps * dc]);
+}
+
+const ADJ_SQUARES: SquareOffset[] = [[0, 1], [1, 1], [1, 0], [1, -1], [0, -1], [-1, -1], [-1, 0], [-1, 1]];
+
+// Function to get all valid moves for the current game state
+export function hasLegalMoves(gameState: FEN): boolean {
+  return CHESSBOARD_SQUARES.some(from => legalMovesFrom(gameState, from).length);
+}
+
+// Function to check if the current player is in checkmate
+export function isCheckmate(gameState: FEN, activeColor: Color): boolean {
+  return isAtomicCheck(gameState.board, activeColor) && !hasLegalMoves(gameState);
+}
+
+// Function to check if the game is in stalemate
+export function isStalemate(gameState: FEN, activeColor: Color): boolean {
+  return !isAtomicCheck(gameState.board, activeColor) && !hasLegalMoves(gameState) && findKing(gameState.board, activeColor) != null;
+}
+
+// Function to check if a pawn at a given square can be promoted
+export function canPromotePawnAt(board: Chessboard, square: Square): boolean {
+  const piece = board[square];
+  if (!piece || PIECE_TO_TYPE[piece] != PieceType.PAWN) return false;
+  return PROMOTION_SQUARES[PIECE_TO_COLOR[piece]].includes(square);
+}
+
+export function legalMovesFrom(gameState: FEN, from: Square): ExternalMove[] {
+  return movesFrom(gameState, from).filter(isLegalMove);
+}
+
+function isLegalMove({ result, color }: ExternalMove): boolean {
+  return isLegalPosition(result, color);
+}
+
+function movesFrom(gameState: FEN, from: Square): ExternalMove[] {
+  return [
+    ...castlesFrom(gameState, CastleSide.KINGSIDE, from),
+    ...castlesFrom(gameState, CastleSide.QUEENSIDE, from),
+    ...doubleMovesFrom(gameState, from),
+    ...enPassantsFrom(gameState, from),
+    ...capturesFrom(gameState, from),
+    ...standardMovesFrom(gameState, from),
+  ];
+}
+
+// Function to get valid standard captures from a given square
+function capturesFrom(gameState: FEN, from: Square): ExternalMove[] {
+  const { board, activeColor } = gameState;
+  const piece = board[from];
+  if (!piece || PIECE_TO_COLOR[piece] != activeColor) return [];
+  const { patterns, steps } = PIECE_CAPTURE_PATTERNS[piece];
+  const enemyColor = getEnemyColor(activeColor);
+  return patterns.flatMap(offset => {
+    for (let i = 1; i <= steps; i++) {
+      const to = moveSquare(from, offset, i);
+      if (!to) return [];
+      const currPiece = board[to];
+      if (currPiece) return (PIECE_TO_COLOR[currPiece] == enemyColor) ? [to] : [];
+    }
+    return [];
+  })
+    .map(to => {
+      const result = structuredClone(board);
+      const move = { from: from, to: to };
+      movePiece(result, move);
+      const explosions = getSurroundingExplosions(result, to);
+      explosions.forEach(square => explodeSquare(result, square));
+      explodeSquare(result, to);
+
+      const flags = [Flag.CAPTURE];
+      [to, ...explosions].forEach(square => {
+        if (board[square] == PieceType.ROOK) {
+          if (from == CASTLE_MOVES[activeColor][CastleSide.KINGSIDE].rookMove.from) {
+            flags.push(Flag.KINGSIDE_ROOK_CAPTURED);
+          } else if (from == CASTLE_MOVES[activeColor][CastleSide.QUEENSIDE].rookMove.from) {
+            flags.push(Flag.QUEENSIDE_ROOK_CAPTURED);
+          }
+        }
+      });
+
+
+      return ({
+        from: from,
+        to: to,
+        piece: piece,
+        flags: flags,
+        color: activeColor,
+        moveType: MoveType.CAPTURE,
+        result: result,
+        actions: [
+          { move: move, explode: true },
+          { move: { from: to, to: to }, explode: true },
+          ...explosions.map(square => ({ move: { from: square, to: square }, explode: true })),
+        ],
+      });
+    });
+}
+
+// Function to get valid standard moves from a given square
+function standardMovesFrom(gameState: FEN, from: Square): ExternalMove[] {
+  const { board, activeColor } = gameState;
+  const piece = board[from];
+  if (!piece || PIECE_TO_COLOR[piece] != activeColor) return [];
+  const { patterns, steps } = PIECE_MOVE_PATTERNS[piece];
+  return patterns.flatMap(offset => {
+    const between: Square[] = [];
+    for (let i = 1; i <= steps; i++) {
+      const to = moveSquare(from, offset, i);
+      if (!to || board[to]) return between;
+      between.push(to);
+    }
+    return between;
+  })
+    .map(to => {
+      const result = structuredClone(board);
+      movePiece(result, { from: from, to: to });
+      const flags = [];
+      if (piece == PieceType.PAWN) {
+        flags.push(Flag.PAWN_MOVE);
+      }
+      if (piece == PieceType.KING) {
+        flags.push(Flag.KING_MOVE);
+      }
+      if (piece == PieceType.ROOK) {
+        if (from == CASTLE_MOVES[activeColor][CastleSide.KINGSIDE].rookMove.from) {
+          flags.push(Flag.KINGSIDE_ROOK_MOVE);
+        } else if (from == CASTLE_MOVES[activeColor][CastleSide.QUEENSIDE].rookMove.from) {
+          flags.push(Flag.QUEENSIDE_ROOK_MOVE);
+        }
+      }
+      if (canPromotePawnAt(result, to)) {
+        flags.push(Flag.PROMOTION);
+      }
+
+      return ({
+        from: from,
+        to: to,
+        piece: piece,
+        flags: flags,
+        color: activeColor,
+        moveType: MoveType.STANDARD_MOVE,
+        result: result,
+        actions: [{ move: { from: from, to: to }, explode: false }],
+      });
+    });
+}
+
+// Function to get valid double moves from a given square
+function doubleMovesFrom({ board, activeColor }: FEN, from: Square): ExternalMove[] {
+  const piece = board[from];
+  if (!piece || PIECE_TO_TYPE[piece] != PieceType.PAWN || PIECE_TO_COLOR[piece] != activeColor) return [];
+  if (!(from in PAWN_DOUBLE_MOVES[activeColor])) return [];
+  const { between, to } = PAWN_DOUBLE_MOVES[activeColor][from];
+  if (board[between] || board[to]) return [];
+
+  const result = structuredClone(board);
+  const move = { from: from, to: to };
+  movePiece(result, move);
+  return [{
+    from: from,
+    to: to,
+    piece: piece,
+    flags: [Flag.DOUBLE],
+    color: activeColor,
+    moveType: MoveType.STANDARD_MOVE,
+    result: result,
+    actions: [{ move: move, explode: false }],
+  }];
+}
+
+// Function to get valid en passant moves from a given square
+function enPassantsFrom({ board, activeColor, enPassantTargets }: FEN, from: Square): ExternalMove[] {
+  const piece = board[from];
+  if (!piece || PIECE_TO_TYPE[piece] != PieceType.PAWN || PIECE_TO_COLOR[piece] != activeColor) return [];
+
+  return PIECE_CAPTURE_PATTERNS[piece].patterns.flatMap(offset => moveSquare(from, offset) ?? [])
+    .filter(to => enPassantTargets.includes(to))
+    .map(to => {
+      const move = { from: from, to: to };
+      const result = structuredClone(board);
+      const r1 = SQUARE_TO_GRID_INDEX[from][0];
+      const c2 = SQUARE_TO_GRID_INDEX[to][1];
+      const target = squareIndexToSquare([r1, c2]);
+      movePiece(result, move);
+      explodeSquare(result, target);
+      const explosions = getSurroundingExplosions(result, to);
+      explosions.forEach(square => explodeSquare(result, square));
+      explodeSquare(result, to);
+
+      return ({
+        from: from,
+        to: to,
+        piece: piece,
+        flags: [Flag.CAPTURE],
+        color: activeColor,
+        moveType: MoveType.EN_PASSANT,
+        result: result,
+        actions: [
+          { move: move, explode: true },
+          { move: { from: target, to: target }, explode: true },
+          ...explosions.map(square => ({ move: { from: square, to: square }, explode: true })),
+        ],
+      });
+    });
+}
+
+function castlesFrom(gameState: FEN, castleSide: CastleSide, from: Square): ExternalMove[] {
+  const { board, activeColor, hasCastlingRights } = gameState;
+  if (!hasCastlingRights[activeColor][castleSide] || isAtomicCheck(board, activeColor)) return [];
+  const { kingMove, between, rookMove } = CASTLE_MOVES[activeColor][castleSide];
+  if (from != kingMove.from) return [];
+  const king = board[kingMove.from];
+  const rook = board[rookMove.from];
+  if (!king || PIECE_TO_TYPE[king] != PieceType.KING || PIECE_TO_COLOR[king] != activeColor) return [];
+  if (!rook || PIECE_TO_TYPE[rook] != PieceType.ROOK || PIECE_TO_COLOR[rook] != activeColor) return [];
+  if (between.some(square => board[square])) return [];
+
+  const result = structuredClone(board);
+  movePiece(result, { from: kingMove.from, to: kingMove.to });
+  movePiece(result, { from: rookMove.from, to: rookMove.to });
+
+  return [{
+    from: from,
+    to: kingMove.to,
+    piece: king,
+    flags: [Flag.KING_MOVE, castleSide == CastleSide.KINGSIDE ? Flag.KINGSIDE_ROOK_MOVE : Flag.QUEENSIDE_ROOK_MOVE],
+    color: activeColor,
+    moveType: castleSide == CastleSide.KINGSIDE ? MoveType.KINGSIDE_CASTLE : MoveType.QUEENSIDE_CASTLE,
+    result: result,
+    actions: [
+      { move: kingMove, explode: false },
+      { move: rookMove, explode: false },
+    ],
+  }];
+}
+
+// Moves a piece from one square to another
+function movePiece(board: Chessboard, { from, to }: Move) {
+  board[to] = board[from];
+  board[from] = null;
+}
+
+// Explodes a piece at a given square
+function explodeSquare(board: Chessboard, square: Square) {
+  board[square] = null;
+}
+
+// Returns a list of surrounding squares that will explode given a target square being captured
+function getSurroundingExplosions(board: Chessboard, square: Square): Square[] {
+  return ADJ_SQUARES.flatMap(offset => moveSquare(square, offset) ?? [])
+    .filter(to => {
+      const piece = board[to];
+      return piece && PIECE_TO_TYPE[piece] != PieceType.PAWN;
+    });
+}
+
+// Finds a specific piece on the board
+export function findPiece(board: Chessboard, piece: Piece): Square | null {
+  return CHESSBOARD_SQUARES.find(key => board[key] == piece) ?? null;
+}
+
+// Finds the king of a specific color on the board
+export function findKing(board: Chessboard, activeColor: Color): Square | null {
+  return findPiece(board, activeColor == Color.WHITE ? 'K' : 'k');
+}
+
+// Checks if two squares are adjacent to each other
+export function isAdjacent(square1: Square, square2: Square) {
+  const [r1, c1] = SQUARE_TO_GRID_INDEX[square1];
+  const [r2, c2] = SQUARE_TO_GRID_INDEX[square2];
+  return Math.abs(r1 - r2) <= 1 && Math.abs(c1 - c2) <= 1;
+}
+
+// Checks if the specified player is in atomic check
+export function isAtomicCheck(board: Chessboard, activeColor: Color): boolean {
+  const enemyColor = getEnemyColor(activeColor);
+  const kingPos = findKing(board, activeColor);
+  const enemyKingPos = findKing(board, enemyColor);
+  if (!kingPos || !enemyKingPos) return false;
+  if (isAdjacent(kingPos, enemyKingPos)) return false;
+  for (let [piece, color] of Object.entries(PIECE_TO_COLOR) as [Piece, Color][]) {
+    if (activeColor == color) continue;
+    const { patterns: pattern, steps } = PIECE_CAPTURE_PATTERNS[piece];
+    for (let offset of pattern) {
+      for (let i = 1; i <= steps; i++) {
+        const square = moveSquare(kingPos, offset, -i);
+        if (!square) break;
+        const piece2 = board[square];
+        if (!piece2) continue;
+        if (PIECE_TO_COLOR[piece2] != enemyColor || PIECE_TO_TYPE[piece2] != PIECE_TO_TYPE[piece]) break;
+        return true;
+      }
+    }
+  }
+  return false;
+}
+
+// Checks if the atomic chess position causes the specified player to lose
+export function isLegalPosition(board: Chessboard, activeColor: Color): boolean {
+  return findKing(board, activeColor) != null && !isAtomicCheck(board, activeColor);
 }
